+++ providers/default.rb
@@ -28,7 +28,7 @@ action :run do
 
     windows_home new_resource.rdp_username do
       password rdp_password
-      sensitive new_resource.sensitive
+      confidential new_resource.confidential
       action :create
     end
 
@@ -47,7 +47,7 @@ action :run do
     template "#{startup_path}/rdp_screenresolution.cmd" do
       source 'rdp_screenresolution.cmd.erb'
       cookbook 'windows_screenresolution'
-      sensitive new_resource.sensitive
+      sensitive new_resource.confidential
       variables(
         target: new_resource.target,
         user: new_resource.username,
@@ -60,28 +60,28 @@ action :run do
     # https://technet.microsoft.com/en-us/library/cc722151%28v=ws.10%29.aspx
     registry_key 'HKLM\SYSTEM\CurrentControlSet\Control\Terminal Server' do
       values [{ name: 'fDenyTSConnections', type: :dword, data: 0 }]
-      sensitive new_resource.sensitive
+      sensitive new_resource.confidential
       action :create
     end
 
     # http://www.mytecbits.com/microsoft/windows/rdp-identity-of-the-remote-computer
     registry_key 'HKLM\SOFTWARE\Microsoft\Terminal Server Client' do
       values [{ name: 'AuthenticationLevelOverride', type: :dword, data: 0 }]
-      sensitive new_resource.sensitive
+      sensitive new_resource.confidential
       action :create
     end
 
     execute 'open rdp firewall' do
       command 'netsh advfirewall firewall add rule name="RDP" protocol=TCP dir=in profile=public'\
       ' localport=3389 remoteip=localsubnet localip=any action=allow'
-      sensitive new_resource.sensitive
+      sensitive new_resource.confidential
       not_if 'netsh advfirewall firewall show rule name="RDP" > nul'
     end
 
     windows_autologin "set #{new_resource.rdp_username}" do
       username new_resource.rdp_username
       password rdp_password
-      sensitive new_resource.sensitive
+      confidential new_resource.confidential
       only_if { new_resource.rdp_autologin }
     end