+++ resources/default.rb
@@ -12,10 +12,4 @@ attribute :rdp_groups, kind_of: Array, default: ['Administrators', 'Remote Deskt
 attribute :rdp_username, kind_of: String, default: 'rdp_local'
 attribute :rdp_password, kind_of: [String, NilClass]
 
-attribute :sensitive, kind_of: [TrueClass, FalseClass] # , default: true - see initialize below
-
-# Chef will override sensitive back to its global value, so set default to true in init
-def initialize(*args)
-  super
-  @sensitive = true
-end
+attribute :confidential, kind_of: [TrueClass, FalseClass], default: true